+++ google-cloud-pubsub/lib/google/cloud/pubsub/subscriber/async_modify_deadliner.rb
@@ -22,9 +22,9 @@ module Google
       class Subscriber
         ##
         # @private
-        # # AsyncDelayer
+        # # AsyncModifyDeadliner
         #
-        class AsyncDelayer
+        class AsyncModifyDeadliner
           include MonitorMixin
 
           attr_reader :batch
@@ -52,7 +52,7 @@ module Google
                   @batch.add deadline, ack_id
                 else
                   unless @batch.try_add deadline, ack_id
-                    publish_batch!
+                    push_batch_request!
 
                     @batch = Batch.new max_bytes: @max_bytes
                     @batch.add deadline, ack_id
@@ -74,7 +74,7 @@ module Google
               break if @stopped
 
               @stopped = true
-              publish_batch!
+              push_batch_request!
               @cond.signal
             end
 
@@ -91,7 +91,7 @@ module Google
 
           def flush
             synchronize do
-              publish_batch!
+              push_batch_request!
               @cond.signal
             end
 
@@ -119,7 +119,7 @@ module Google
                 time_since_first_publish = Time.now - @batch_created_at
                 if time_since_first_publish > @interval
                   # interval met, publish the batch...
-                  publish_batch!
+                  push_batch_request!
                   @cond.wait
                 else
                   # still waiting for the interval to publish the batch...
@@ -129,7 +129,7 @@ module Google
             end
           end
 
-          def publish_batch!
+          def push_batch_request!
             return if @batch.nil?
 
             request = @batch.request
